notebook upload
notebook contains the preprocessing with tokenization and embedding using Finbert. Then there is a creation and population of ChromaDB with the relative embedding for each sentence. After that there is a formulation of 7 queries with the most relevant 3 sentences in terms of distance with the respect of the query provided (and other parameters). In the final part, actually, is placed a sentiment analysis based on "mrm8488/distilroberta-finetuned-financial-news-sentiment-analysis" that provided the sentiment of each resulting document in real time and, after that, there is a classification report with confusion matrix in respect of ground truth.